Add files via upload
This system is designed to provide a secure
environment where administrators can access and manipulate critical data through a
password-protected portal. After successful login, the administrator is directed to the
homepage that serves as the central hub for navigating to different sections of the hospital's
database, including entities such as doctors, patients, staff, and appointments. The site
includes comprehensive functionalities that allow for the creation, reading, updating, and
deletion (CRUD) of information, which is important for keeping accurate and up-to-date
records. Furthermore, we have included advanced search functionality that allows users to
find specific data related to any entered name, which improves the database system's
efficiency. This project aims to improve healthcare service delivery by optimizing hospital
administrative tasks and implementing effective data management